Move  into it's own file. Cleanup require statements
